api/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile /tmp/tmp.be_skeleton_builder/api/requirements/dev.in
#
appdirs==1.4.3            # via black
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, packaging, pytest
backcall==0.1.0           # via ipython
bandit==1.6.2
black==19.3b0
click==7.0                # via black, pip-tools
coverage==4.5.4           # via pytest-cov
decorator==4.4.0          # via ipython, traitlets
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django==2.2.16
djangorestframework==3.10.3
gitdb2==2.0.5             # via gitpython
gitpython==3.0.2          # via bandit
gunicorn==19.9.0
importlib-metadata==0.19  # via pluggy, pytest
ipdb==0.12.2
ipython-genutils==0.2.0   # via traitlets
ipython==7.7.0
jedi==0.15.1              # via ipython
mccabe==0.6.1             # via pylama
more-itertools==7.2.0     # via pytest, zipp
packaging==19.1           # via pytest
parso==0.5.1              # via jedi
pbr==5.4.2                # via stevedore
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==4.1.0
pluggy==0.12.0            # via pytest
prompt-toolkit==2.0.9     # via ipython
psycopg2-binary==2.8.3
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via pylama
pydocstyle==4.0.1         # via pylama
pyflakes==2.1.1           # via pylama
pygments==2.4.2           # via ipython
pylama==7.7.1
pyparsing==2.4.2          # via packaging
pytest-bandit==0.5.2
pytest-cov==2.7.1
pytest-django==3.5.1
pytest==5.1.1
pytz==2019.2              # via django
pyyaml==5.1.2             # via bandit
six==1.12.0               # via bandit, django-extensions, packaging, pip-tools, prompt-toolkit, stevedore, traitlets
smmap2==2.0.5             # via gitdb2
snowballstemmer==1.9.0    # via pydocstyle
sqlparse==0.3.0           # via django, django-debug-toolbar
stevedore==1.30.1         # via bandit
toml==0.10.0              # via black
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit, pytest
werkzeug==0.15.5
zipp==0.6.0               # via importlib-metadata
pyowm==3.1.1
django-filter==2.4.0
djangorestframework-csv==2.1.0
vcrpy-unittest==0.1.7
pre-commit==2.9.3

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via ipdb, ipython
